relocate templates folder
