+++ modules/utils.js
@@ -1,6 +1,6 @@
 import Components from './components/components.js';
-import blankTemplate from './components/templates/blank.template.js';
-import landingTemplate from './components/templates/landing.template.js';
+import blankTemplate from './templates/blank.template.js';
+import landingTemplate from './templates/landing.template.js';
 
 const EVENTS = {
   TOUCH_MOVE: "touchmove",
@@ -141,7 +141,7 @@ function renderDraggableElements(containerElementID) {
 function renderTemplateIcons() {
   const templateContainer = document.querySelector('#template-container');
   const templatesList = [landingTemplate, blankTemplate];
-  console.log(templatesList);
+  
   for (const template of templatesList) {
     
     const container = document.createElement('div');
